# server/docs/swagger.yaml
swagger: "2.0"
info:
  description: "AgriDealer API"
  version: 0.1.0
  title: "AgriDealer"
  # termsOfService: "http://swagger.io/terms/"
  contact:
    email: "sourabh.chakraborty@gmail.com"
  license:
    name: "Private License"
    # url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/api"
...
schemes:
# - "https"
- "http"
produces: 
- "application/json"
paths:
  /users/{id}:
    parameters:
      - name: "id"
        in: "path"
        description: "User id"
        required: true
        type: "integer"
    post:
      tags:
      - "user"
      summary: "Create user"
      description: "Get use details.  Only same user, or admin users can do this"
      responses:
        default:
          description: "successful operation"
      security: 
      - api_key: []
    patch:
      tags:
      - "user"
      summary: "Update user"
      description: "Update a user.  Only same user or admin can do this"
      responses:
        default:
          description: "successful operation"
      parameters:
      - in: "body"
        name: "body"
        description: "User update object"
        required: true
        schema:
          $ref: "#/definitions/User"
      security: 
      - api_key: []
...
securityDefinitions:
  api_key:
    type: "apiKey"
    name: "x-token"
    in: "header"
definitions:
...
  User:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
      username:
        type: "string"
      firstName:
        type: "string"
      lastName:
        type: "string"
      email:
        type: "string"
      password:
        type: "string"
      phone:
        type: "string"
      userStatus:
        type: "integer"
        format: "int32"
        description: "User Status"
...